backend/requirements.txt
Flask
flask-cors
pymysql
python-dotenv
